use ClassBytecodeWriter.BYTECODE_VERSION for the jvm version of Block return types
ClassBytecodeWriter determines the current running jvm's version and uses that as the target. BlockReturnTypeCreator had 1.7 as its hard coded version which was breaking the jdk 1.6 build. This fixes the openjdk 1.6 build by using the runtime jvm version.

It would be better to have a way to set the target version explicitly, then we could build for 1.5-6 from 1.7.